feat: add user module

+++ backend/src/modules/user/use-cases/delete/user-delete.use-case.ts
@@ -1,21 +1,21 @@
 import { Injectable } from '@nestjs/common';
 import { IUserRepository } from '../../repositories/iuser.repository';
-import { UserEntity } from 'src/domain/entities';
 import { IBaseUseCase } from 'src/common/interfaces';
-import { UserFindOneByIdUseCase } from '../find-one-by-id/user-find-one-by-id.use-case';
+import { ResourceNotFoundError } from 'src/domain/errors/resource-not-found';
 
 @Injectable()
-export class UserDeleteUseCase implements IBaseUseCase<UserEntity, string> {
+export class UserDeleteUseCase implements IBaseUseCase<void, string> {
   constructor(
-    private readonly userRepository: IUserRepository,
-    private readonly userFindOneByIdUseCase: UserFindOneByIdUseCase,
+    private readonly userRepository: IUserRepository
   ) {}
 
-  async execute(id: string): Promise<UserEntity> {
-    await this.userFindOneByIdUseCase.execute(id);
+  async execute(id: string): Promise<void> {
+    const user = await this.userRepository.findById(id);
 
-    const user = await this.userRepository.delete(id);
+    if (!user) {
+      throw new ResourceNotFoundError();
+    }
 
-    return user;
+    await this.userRepository.delete(id);
   }
 }
